requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
_tflow_select=2.3.0=mkl
absl-py=0.10.0=py37hc8dfbb8_1
astor=0.8.1=pyh9f0ad1d_0
audioread=2.1.8=py37hc8dfbb8_3
brotlipy=0.7.0=py37hb5d75c8_1001
bzip2=1.0.8=h516909a_3
c-ares=1.16.1=h516909a_3
ca-certificates=2020.6.20=hecda079_0
cached-property=1.5.1=py_0
certifi=2020.6.20=py37he5f6b98_2
cffi=1.14.3=py37h00ebd2e_1
chardet=3.0.4=py37he5f6b98_1008
cryptography=3.2.1=py37hc72a4ac_0
cycler=0.10.0=py_2
dataclasses=0.6=pypi_0
decorator=4.4.2=py_0
ffmpeg=4.3.1=h3215721_1
ffmpeg-python=0.2.0=py_0
filetype=1.0.7=pypi_0
freetype=2.10.4=h7ca028e_0
future=0.18.2=py37hc8dfbb8_2
gast=0.2.2=py_0
gettext=0.19.8.1=hf34092f_1004
gmp=6.2.0=h58526e2_4
gnutls=3.6.13=h79a8f9a_0
google-pasta=0.2.0=pyh8c360ce_0
grpcio=1.33.2=py37haffed2e_2
h5py=3.1.0=nompi_py37h1e651dc_100
hdf5=1.10.6=nompi_h1022a3e_1110
idna=2.10=pyh9f0ad1d_0
imagehash=4.1.0=pypi_0
imageio=2.9.0=pypi_0
importlib-metadata=2.0.0=py_1
joblib=0.17.0=py_0
jpeg=9d=h36c2ea0_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.0=py_0
kiwisolver=1.3.1=py37hc928c03_0
krb5=1.17.1=hfafb76e_3
lame=3.100=h14c3975_1001
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.35=h769bd43_9
libblas=3.9.0=2_openblas
libcblas=3.9.0=2_openblas
libcurl=7.71.1=hcdd3856_8
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.2.1=he1b5a44_1007
libflac=1.3.3=he1b5a44_0
libgcc-ng=9.3.0=h5dbcf3e_17
libgfortran-ng=9.3.0=he4bcb1c_17
libgfortran5=9.3.0=he4bcb1c_17
libgomp=9.3.0=h5dbcf3e_17
libiconv=1.16=h516909a_0
liblapack=3.9.0=2_openblas
libllvm8=8.0.1=hc9558a2_0
libnghttp2=1.41.0=h8cfc5f6_2
libogg=1.3.2=h516909a_1002
libopenblas=0.3.12=pthreads_h4812303_1
libpng=1.6.37=h21135ba_2
libprotobuf=3.13.0.1=h8b12597_0
librosa=0.7.2=py_1
libsndfile=1.0.29=he1b5a44_0
libssh2=1.9.0=hab1572f_5
libstdcxx-ng=9.3.0=h2ae2ef3_17
libtiff=4.1.0=h4f3a223_6
libvorbis=1.3.7=he1b5a44_0
libwebp-base=1.1.0=h36c2ea0_3
llvmlite=0.31.0=py37h5202443_1
lz4-c=1.9.2=he1b5a44_3
markdown=3.3.3=pyh9f0ad1d_0
matplotlib-base=3.3.2=py37hc9afd2a_1
mir-eval=0.6=pypi_0
ncurses=6.2=h58526e2_3
nettle=3.4.1=h1bed415_1002
networkx=2.5=pypi_0
norbert=0.2.1=py_0
numba=0.48.0=py37hb3f55d8_0
numpy=1.19.4=py37h7e9df27_0
olefile=0.46=pyh9f0ad1d_1
openh264=2.1.1=h8b12597_0
openssl=1.1.1h=h516909a_0
opt-einsum=3.3.0=pypi_0
packaging=20.4=pypi_0
pandas=0.25.1=py37hb3f55d8_0
pillow=8.0.1=py37h63a5d19_0
pip=20.2.4=py_0
pooch=1.2.0=pypi_0
protobuf=3.13.0.1=py37h3340039_1
pycparser=2.20=pyh9f0ad1d_2
pymongo=3.11.0=pypi_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyrubberband=0.3.0=pypi_0
pysndfx=0.3.6=pypi_0
pysocks=1.7.1=py37he5f6b98_2
pysoundfile=0.9.0.post1=pypi_0
python=3.7.8=h6f2ec95_1_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytz=2020.4=pyhd8ed1ab_0
pywavelets=1.1.1=pypi_0
readline=8.0=he28a2e2_2
requests=2.24.0=pyh9f0ad1d_0
resampy=0.2.2=py_0
scikit-image=0.17.2=pypi_0
scikit-learn=0.23.2=py37hddcf8d6_2
scipy=1.5.3=py37h14a347d_0
setuptools=49.6.0=py37he5f6b98_2
six=1.15.0=pyh9f0ad1d_0
spleeter=1.5.3=py37hc8dfbb8_2
sqlite=3.33.0=h4cf870e_1
tensorboard=1.15.0=py37_0
tensorflow=1.15.2=pypi_0
tensorflow-estimator=1.15.1=pyh2649769_0
termcolor=1.1.0=py_2
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2020.10.1=pypi_0
tk=8.6.10=hed695b0_1
torch=1.7.0=pypi_0
tornado=6.1=py37h4abf009_0
typing-extensions=3.7.4.3=pypi_0
urllib3=1.25.11=py_0
werkzeug=0.16.1=py_0
wheel=0.35.1=pyh9f0ad1d_0
wrapt=1.12.1=py37h8f50634_1
x264=1!152.20180806=h14c3975_0
xz=5.2.5=h516909a_1
zipp=3.4.0=py_0
zlib=1.2.11=h516909a_1010
zstd=1.4.5=h6597ccf_2
